+++ keyboards/htx_studio/.vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "one_handed_keyboard.h": "c"
+    }
+}
