+++ keyboards/htx_studio/one_handed_keyboard/halconf.h
