+++ keyboards/htx_studio/one_handed_keyboard/info.json
@@ -1,0 +1,25 @@
+{
+    "manufacturer": "HTXStudio",
+    "maintainer": "HTXStudio",
+    "bootloader": "stm32-dfu",
+    "diode_direction": "ROW2COL",
+    "encoder": {
+        "rotary": [
+            {"pin_a": "A9", "pin_b": "A10"}
+        ]
+    },
+    "features": {
+        "bootmagic": true,
+        "encoder": true,
+        "extrakey": true,
+        "mousekey": true,
+        "nkro": true,
+        "pointing_device": true
+    },
+    "matrix_pins": {
+        "cols": ["A0", "A1", "A2", "C4", "B0", "B1", "B2", "B7", "B5", "B3", "B10", "C11", "C10", "A15"],
+        "rows": ["B12", "B15", "C6", "B13", "B14", "A8"]
+    },
+    "processor": "STM32G431",
+    "url": "https://github.com/htx-studio"
+}
