+++ keyboards/htx_studio/one_handed_keyboard/mcuconf.h
