+++ keyboards/htx_studio/one_handed_keyboard/ml/readme.md
@@ -1,22 +1,22 @@
-# One-Handed_Keyboard_MR
+# One-Handed_Keyboard_ML
 
-![右手大键盘](https://imgur.com/y5O0E2E.jpg)
+![左手大键盘](https://imgur.com/QjQSxhP.jpg)
 
 This is a keyboard with a fan-shaped layout and 59 keys.The commonly used buttons are placed near the thumb.
 
 * Keyboard Maintainer：[HTX Studio](https://github.com/htx-studio)
-* Hardware Supported：HTX_STUDIO_ONE-HANDEN_KEYBOARD_MR  PCB
+* Hardware Supported：HTX_STUDIO_ONE-HANDEN_KEYBOARD_ML  PCB
 * Hardware Availability：[One-Handed-Keyboard](https://github.com/htx-studio/One-Handed-Keyboard/tree/main/Hardware)
 
 ## How to compile
 
 Make example for this keyboard (after setting up your build environment):
 
-    `make htx_studio/one_handed_keyboard_mr:default`
+    `make htx_studio/one_handed_keyboard/ml:default`
 
 Flashing example for this keyboard:
 
-    `make htx_studio/one_handed_keyboard_mr:default:flash`
+    `make htx_studio/one_handed_keyboard/ml:default:flash`
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
