+++ keyboards/htx_studio/one_handed_keyboard/sl/keymaps/default/rules.mk
