--- keyboards/htx_studio/one_handed_keyboard_sl/config.h
@@ -1,30 +1,0 @@
-/* Copyright 2025 htx-studio (@https://github.com/htx-studio)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-#define SPI_DRIVER SPID1
-#define SPI_SCK_PIN A5
-#define SPI_MISO_PIN A6
-#define SPI_MOSI_PIN A7
-
-#define POINTING_DEVICE_CS_PIN A4
-#define POINTING_DEVICE_MOTION_PIN A3
-#define POINTING_DEVICE_TASK_THROTTLE_MS 1
-#define POINTING_DEVICE_INVERT_X 1
-
-#define ENCODER_RESOLUTION 2
-
